refactor(core): home pr_assignee in core/merge and pin fast_push as a root leaf
The flat-core ratchet caught both new leaves: pr_assignee joins its
pr_create_verify/pr_slug_resolution siblings in core/merge/; fast_push
stays a root leaf with the pin raised 66->67 (a whole ship-flow
alternative owned by no existing subpackage).

Open questions & assumptions:
- none
